requirements.txt
attrdict==2.0.1
common==0.1.2
joblib==1.1.0
lightning_utilities==0.8.0
matplotlib==3.5.1
numpy==1.21.5
openai==0.27.8
optuna==3.1.0
pandas==1.4.2
pytorch_lightning==1.9.4
pytorch_metric_learning==0.9.81
PyYAML==6.0
scikit_learn==1.0.2
scipy==1.7.3
tiktoken==0.4.0
torch==1.12.1+cu116
torchmetrics==0.11.0
tqdm==4.64.0
transformers==4.22.2
